Add files via upload
Adding TFCmetals, TFC Pewer and TFC Wells as a counter to the disabling of infinite water sources. TFCmetals is a dependency for TFC pewter which is an alloy and is used for utensils such as non-clay water jugs. TFC Wells is a block that adds the well functionality that lets you create a more reliable water source than rainbarrels. From https://github.com/Peffern/Mod-Jars